Add the 'RunBuilder'
This structure bundles the three objects necessary to define the
runner's journey: the sprints, the eaten cells, and the journey
itself. Each of these objects depends on the other so defining
them independently would be difficult.

The culmination of this change is the 'make_run', function which
constructs the RunBuilder corresponding to the runner's journey
of 'n' steps.

This change also fixes an earlier mistake where the maximum length
of a sprint was limited to '2p' rather than '2p + 1'