Theorems on 'make_run' and 'run_path'
These theorems are focused on basic manipulation of 'make_run':
 - Identifying the base case of a 'make_run' component
 - Expanding the recurrence of a 'make_run' component
 - Simplifying the length of a 'make_run' component
 - Rewriting part of a 'make_run' in terms of a shorter one

The change also proves a relationship between the cells of a runner
journey and the final location in the corresponding sprint

Lastly, the change defines a 'RunPath', which is a flattened list
of all the RunStates in the sprints of a 'make_run'. It proves that
every RunState in the RunPath must appear as 'sprints[i][j]' for
some i and j.