Prove 'make_run_runner_avoids_eaten_cells_of_nice'
This result states that if the devil is "nice", the runner will never
visit a cell that was previously eaten. This result handles one of the
three cases we need to prove for 'run_path_not_eaten_of_nice'
(the more general result that the RunPath and eaten list have no
cells in common).