Add Trace.lean
A 'trace' is similar to our current definition of a 'sprint' but
without restrictions on length, and starting location. Traces also
aren't limited to positive x-coordinates. The idea is that sprints
will be redefined in terms of traces and the wall along the y-axis
will be reintroduced by adding blocking cells (in addition to the
'eaten' cells).

This approach will allow us to prove that the 'RunPath' is equal to a
particular trace. Then we can use theorems on traces to complete the
final proof. The refactor is necessary because some of the logic in
the original paper considers a path which enters quadrant #2 of the
plane. This is disallowed by our current definitions but not by the
definition of a trace.