Add Sprint.lean and update Runner.lean to match
This completes the trace/sprint refactoring. Previously, sprints were
defined in Runner.lean and included logic for following the y-axis
and stopping before the angel's maximum range was exceeded. However,
the final steps of Mathe's proof require us to consider the runner's
behavior when one or both of these limits is removed.

With the refactored code, a "trace" is the path taken by the runner
for a given starting point, length, and list of blocked cells.
A "sprint" is then defined as a trace which respects the angel's range.
Finally, the function "make_run" assembles a sequence of sprints which
begins at (0, 0) and defines the list of blocked cells such that the
runner never steps on an eaten cell or crosses the y-axis.