Add Region.lean and Dupes.lean
The results contained in the former will be used to reason about
regions of orthogonally connected cells in a grid. To represent
such a region we will use a List, bundled with an assertion that the
list contains no duplicates. The results needed for reasoning about
duplicates in a list are contained in Dupes.lean

Note that Dupes.lean is expected to be mostly complete at this point
while Region.lean only has the definition of a 'RegionBuilder' with a
few basic operations. These will eventually be used to implement the
algorithm for building a region.