Fix pervasive off-by-one error in sprint length
I had previously defined 'sprint' so that the maximum length was 2p.
After carefully reviewing the original paper I discovered that I had
made an off-by-one error and that the actual maximum length should be
set at 2p+1.

The confusion comes from the fact that (in the relevant section of the
paper) 'n' refers to the *index* of the last element of the sprint, not
the *length* of the sprint. So when the paper states that n <= 2p it is
easy to misread that as setting the maximum sprint length to 2p, rather
than 2p+1.

With the previous definition it was possible to have a sprint of zero
length, but not with the new definition. Previously, many (nearly all)
theorems in Sprint.lean and Runner.lean needed to either prove that a
sprint was non-nil, or take a statement to that effect as an argument.
Since all sprints are non-nil now, we can remove these requirements.
The upside is that this should simplify things going forward. The
downside is that this change is much longer than it otherwise would be
as a result!