Prove 'run_path_length_increasing_of_ppos'
That is, the length of the run path increases with the number of
sprints. To simplify this proof, we also showed that any sprint in
a 'make_run' can be constructed by calling "sprint" with some
arguments. This allows us to use our existing results on sprints,
in particular 'sprint_length_lb'.