Refactor 'make_run_next_sprint_avoids_blocked...'
The logic in the induction step of this proof is likely required for
other proofs also. Factor it out and generalize it into its own proof.
Specifically, this is the logic that shows a sprint can never reach a
far cell in the west wall and will never have a cell eaten by a nice
devil on some future step.