Prove 'trace_split'
This result shows how we can write a trace as two appended traces.
This change also includes a proof of an alternate, possibly more useful
form of 'trace_getElem_recurrence'. Though I didn't use it in this
proof, it may be useful in the future so I decided to keep it.