Prove 'run_path_eq_trace_of_nice'
This is *the* main result of Runner.lean. It states that any 'RunPath'
is equivalent to a particular trace.

The final step of the paper shows an incompatibility between the
RunPath of a trapped angel and a trace which circumnavigates a
particular contiguous region of cells. That argument is only possible
if the former can be shown to be equal to a segment of the latter.
This result demonstrates that equivalence.