Add the path constraint to RegionBuilder
This change is a significant expansion / rewrite of Region.lean and
proves the defining characteristic of a region. That is, if a path
exists from a given cell to the region start, that cell will be in
the region.

Note that the constaint actually proven is slightly different because
it needs to also hold for partially constructed regions. In particular
it states that either the given cell is in the region (as described
above), or a path exists from the given cell to some pending cell.
For completed regions this second option will drop out because the
pending list is empty.

The next step will be to propagate this result up to 'make_region_impl'
and 'Region'