Switch Region from type 'List' to type 'Finset'
Up to this point our theorems on regions have used the predicate
'list_mem_iff' to indicate that two lists are mutual subsets.
Unfortunately this is inadequate for proving some of the key results on
regions - we need actual equality. The easiest way to achieve this is
to switch from List to Finset.

This change also removes some unused code and changes the definition
of 'hstart' to be more precise. In particular, rather than stating that
start is in 'region' or 'pending', now start is required to be at the
end of 'region' or the beginning of 'pending'. This precision will be
required to prove the "region recurrence" - the next major result.