Prove 'region_recurrence_exists'
This result states that for any region with more than one cell there
is some cell we can remove such that the result is still a region.
This will primarily be used to prove 'region_edges_le', which puts an
upper bound on the length of the perimeter of a region based on its
size.