Add Perimeter.lean
This file defines the concept of an "edge", which includes one cell in
a set, and an adjacent edge not in that set. It defines a function for
constructing the set of edges for a region (that is, the perimeter)
and proves that these edges have the desired properties.

Most significantly, this file proves 'region_edges_le', which places
an upper bound on the number of edges of a region based on its size:
A region of size 'n' will have no more than 2n+2 edges.