Prove 'edge_of_runner_mem_region_edges'
This result states that every edge to the runner's left appears in the
perimeter of the region of blocked cells. This will be used together
with 'region_edges_le' to put an upper bound on the length of the
runner's path as a function of the number of blocked cells.