Add Unit.lean
This file introduces the 'UVec' type for unit vectors. It also defines
the finite set of all unit vectors and proves the set is complete. It
proves that the cardinality of the Finset is '4'. This will be used to
prove the cardinality of the set of RunState within the escape square
which will be needed to show that the runner's path eventually loops.

We make use of unit vectors in Perimeter.lean (to define edges),
Region.lean (to define adjacency), and in Trace.lean (in the definition
of the RunState). Previously this was done on an ad-hoc basis, leading
to complicated and / or duplicated proofs. With this change we can
also clean up relevant definitions and results in those files.