Prove 'run_path_loops_of_nice_devil_wins'
That is, if the nice devil wins, the runner's path eventually loops.
In order to prove this we demonstrate an equivalence between the
run states inside the escape square and the cartesian product of three
finite sets. Since that means the set of close run states is finite we
can use the pigeonhole principle to complete the proof.