Add list_has_dupes and duplicate decidability
Previously we were using the negation of 'list_nodupes' to indicate
that a list *did* have duplicates. Now we have 'list_has_dupes'.
Also, this makes 'list_nodupes' and 'list_has_dupes' decidable.
This was done by introducing an alternate definition -
'list_nodupes_decidable' - which is trivially decidable and proving
this is equivalent to the original definition.