Add 'find_first_dupe' to Dupes.lean
This function finds the location of the first element in a list such
that some earlier element has the same value. Note that this is
different from the (possibly) more intuitive definition: to find the
first element which is repeated anywhere in the list.

The change also proves that 'find_first_dupe' does in-fact find a
duplicate value and that it is in-fact the "first" such value
(under the definition described above).