Prove 'trace_start_is_first_dupe'
That is, if a trace loops and is "valid", the first repeated state will
be the initial state. In order to complete the proof, this change also
defines 'undo_next_step' and proves the conditions under which
'next_step' and 'undo_next_step' are inverses.

Note that this proof was also the motivation for adding
'find_first_dupe' to Dupes.lean (in a previous change) since that
function is used in this theorem's statement.