Improved, improved 'find_first' and 'find_last'
Previously these functions required the predicate to take an argument
of type 'Fin (n + 1)'. This made them very awkward to use. The new
implementations don't have this restriction. That is, the argument
can be 'Fin n' for any non-zero n.

Note that this change should allow the client code to be simplified in
many places, but should also be a drop-in replacement. I made some
improvements to Dupes.lean, but all other client code has been left
mostly untouched.

Also note that this is the third version of these functions I have
implemented. Hopefully it will be the last!