Prove 'make_run_x_nonneg'
That is, the runner never goes west of the y-axis. In order to prove
this we needed to extend the west wall below the x-axis as well.
That means we'll have to remove those cells again as part of the final
step of the proof.

We also needed to prove new versions of 'trace_avoids_blocked' and
'sprint_avoids_blocked' and also some new results for the case where
p = 0. There were also some simplifications made to existing proofs to
take advantage of the new results.