Prove 'run_path_intersects_posx_of_south'
To quote the inline comment:
If the runner ever wanders south of the x-axis, there must be some step
along their path where they are on the positive x-axis, facing south or
west.

This result will be used to setup the "endgame" for the proof: that
if the devil wins, there is some prefix of the runner's path which
corresponds to a duplicate-free trace whose final state is on the
x-axis facing south.