Prove 'run_start_valid_of_nice'
That is, the normal starting state for the runner (on the origin,
facing north) is a "valid" starting cell for any nice devil. The
origin itself is not blocked, and the cell to the runner's left
(-1, 0) *is* blocked. The former is true because a nice devil
never eats the origin, and the latter is true because the west wall
will always include (-1, 0).