Prove 'run_path_start_repeats_has_earlier_sfyzxnn'
That is, if a run path revisits the start, it must contain a south
south-facing step with y = 0 and non-negative x. The body of this proof
is similar to the old theorem ('run_path_xaxis_south_of_path_loops')
which is now proven as a consequence of this.

The reason for this reorganization is that the old theorem says
nothing about the relative order of the 'sfyzxnn' step, and the step
which revisits the start. In order to prove that the 'Endgame' has no
duplicates, we need to be sure that the 'sfyzxnn' step comes first.