Define and construct the Endgame
This structure describes the segment of a run path in which a
"nice" devil is able to force the runner back to the x-axis. Recall
that this is a requirement for the devil to win the game. The final
step of the proof will be to show that such an Endgame cannot exist
for 1 < p.

Also bundled in the Endgame structure are proofs that this segment
has no duplicates, never crosses below the x-axis, never faces west
when on the x-axis, and is based on the shortest run path which
crosses the x-axis.