Prove cases 1 and 2 for 'segment_length_lb'
That is, prove that the lower bound on segment length holds for
segments of length 1, and follows from induction for segments that
begin with a forward step. I expect there to be 6 cases in-total.
Here are the other four:
 3) Segment begins with a right turn
 4) Segment begins with a left turn and contains no other turns
 5) Segment begins with a left turn, moves straight for some number
    of steps, and then turns left again
 6) Segment begins with a left turn, moves straight for some number
    of steps, and then turns right

This change also includes some useful identities relating the start
and end of split segments to the original segment and to each-other.
It also includes proofs for the lengths of split segments.