Prove 'segment_length_lb_case3'
That is, prove the segment length lower bound for 'L' segments.
L segments are those which begin with a left turn and then proceed
straight for the remainder of the segment.

This change also includes the proof of 'segment_end_of_L' which
specifies the last RunState of an L segment in terms of the starting
segment. This results will be key in proving the next two cases
in the segment length lower bound: S-segments and U-segments.