Change how TraceSegment splitting is done
Previously, the argument 'k' given to indicate the split location was
relative to the full trace. With this change, it is relative to the
beginning of the segment instead. As a result, we no longer have to
reference seg.i and seg.j (the starting and ending locations of the
trace, which are intended to be private) and can instead reference only
the segment length.