Prove the new 'segment_length_lb_case3'
This corresponds to proving the induction step of the segment length
lower bound for segments that begin with a right turn. Note that
previously this name referred the 'L-segment' case which has now been
renamed 'segment_length_lb_case4'. This was the original intended
numbering of the cases, I just skipped one!