Prove 'segment_length_lb_case5'
That is, prove the induction step of the segment length lower bound
for segments which begin with a 'U-segment'. That is, for segments
which begin with a left turn, some (possibly zero) number of forward
movements, followed by a left turn, and some (possibly zero) number of
additional steps.

To complete the segment length lower bound proof I still need to define
'S-segments' and prove the induction step for segments which begin
with a prefix of that type. That will be case 6 of the proof. Then the
six cases will be combined into the final proof of the segment length
lower bound.