Prove 'segment_length_lb_case6'
That is, prove the induction step for the proof on the segment length
lower bound when the segment begins with an 'S-segment'. Now that the
proofs of all six cases are complete, the final segment length lower
bound proof can be assembled.

Also, I discovered that "omega" can be used to simplify the arithmetic
and algebra in some of the earlier proofs in this file (case 4 in
particular) and have updated them accordingly.