Prove 'trace_trim_neg_one_unchanged'
To quote the inline comment:
Prove that using 'blocked_trim_neg_one' to trim the blocked list has no
effect on the trace if all steps in the trace have non-negative
y-coordinate and never face west when the y-coordinate is zero.

This proof is a bit tedious as it involves considering each of the
three starting directions together with each of the three possible
steps to show that each case either leads to a prohibited state or
behaves the same for both blocked lists.