Define 'blocked_trim_quad1'
Also, prove the conditions under which a trace will remain unchanged
when applying this filter to its blocked list. Note that this is the
filter which will be used to create the final blocked list, with
'top' set to match the y-coordinate of the top cell in the west wall.