Prove 'endgame_trace_eq' and 'endgame_endpoint_lt'
The first states that the endgame trace is equal to the trace with
length (E.i + 1) that begins at 'run_start' and uses the endgame
blocked list. The second states that the index of the "endpoint" of
the endgame trace (that is, 'E.i' for a given endgame 'E') is less
than the length of the perimeter trace.

This change also defines the "perimeter trace" as the longest trace
with no duplicates that begins at run_start and uses the endgame
blocked list.