Prove the "endpoint" and "southwest point"
That is, prove the last point in the endgame trace has y-coordinate of
zero and a non-negative x-coordinate. Also, prove that the last
point in the perimeter trace has location (-1, -1).

This change also includes 'endgame_endpoint_lb' which proves the
lower bound on the length of the endgame trace based on the minimum
length of a sprint.