Prove 'endgame_west_point_eq'
That is, the second-to-last step in the perimeter trace has coordinates
(-2, 0) and is facing south. Note that the corresponding proofs for
the endpoint, southwest point, and southeast point had previously only
included the location of those steps (saying nothing about the
direction). Since that information will definitely be needed for
certain proofs (such as this one) this change also updates those proofs
accordintly.

This change also includes several supporting lemmas to make it easier
to work with the perimeter trace and a new proof of non-membership
for the final blocked list (for -1 < x)