Create 'Endgame.lean'
This file will contain most of the endgame-related results that were
previously proven in Main.lean. These don't rely directly on whether
the devil wins or loses (as the results in Mean.lean do) but depend
only on the constraints established in the definition of Endgame.

The main result of this new file will be to prove a lower bound on the
length of the endgame perimeter. Then that result will be used in
Main.lean to show that the angel of power two wins.