Prove 'endgame_west_of_wall'
That is, prove that a cell 'i+1' steps from the end of the perimeter
trace has coordinates (-2, i) and is facing south, for i less than
the height of the west wall.

Note that the west wall was originally a fair bit taller than intended
but will be trimmed to the intended height by the "quad1" filter. To
clarify, the blocked list used to create the endgame blocklist is
later than it needs to be and so will include a couple extra devil
moves and a taller wall (at least, prior to the trimming) but none of
this will change the logic in the final proof.