Prove 'endgame_north_point_eq'
That is, prove the coordinates and direction for the perimeter cell
located at the top of the west wall. Specifically, the cell at index
P - np - 4 has coordinates (-1, np + 2) and is facing west.

Note that while working on this proof, a mistake was discovered in
'endgame_west_of_wall' which is corrected in this change. Note that
the statement proven previously wasn't false (which is impossible)
but it only applied to cells up to y = np, while the top-most
west-of-wall perimeter cell is (-2, np + 1).