Define 'endgame_wall_height'
Previously we were using hard-coded values for the wall height, but
that makes the code brittle and difficult to change if we need to make
the wall a different height. Now "endgame_wall_height" is opaque to
most of the endgame code.

Note that getting the wall height exactly right, and getting it to
match the wall height in the last 'make_run' block list has been
non-trivial. Fortunately, there is no problem using a larger wall, but
it must be done in such a way that we can prove a tight upper bound for
the size of the endgame blocked list. So if we trim some cells off the
top of the wall (as we were previously doing) the bound may be more
difficult to prove.

For now the wall height will be p(n + 2) + 2.

As an amusing aside: I can understand now why the original paper's
author just added an arbitrary 100 cells to the top of the wall!
It was as if to say, "Just assume it's tall enough!".