Prove 'endgame_southern_border'
That is, prove that the cells between the "endpoint" and the east point
on the perimeter are all of the form

    ((endgame_endpoint E).x + 1 + i, -1, uvec_right)

This change also includes the definition of the index of the
east point and some basic results concerning it. Next will be to prove
that the east point has y-coordinate zero and an x-coordinate at least
two greater than that of the 'endpoint'.