Prove 2 <= x and y = 0 for the endgame east point
This was the last of the six critical points whose attributes needed
to be proven. The next step in the proof will be to use the lower
bound from Segment.lean to prove a lower bound for the length of the
endgame perimeter. Before this can be done, there is one more step:
proving that the index of the east point is less than the index of the
north point.