Prove 'endgame_perimeter_length_lb'
That is, the lower bound for the length of the perimeter is given by
p(t-1) + 7 + 2N, where 'p' is the angel's power, 't' is the number of
turns elapsed before the angel is forced to the x-axis, and 'N' is the
height of the west wall. This is the primary result of Endgame.lean

Note that the bound proven here is actually greater by one than the
bound proven in the original paper. For more information as to why that
is, see the lengthy comment which precedes the proof.