Fix 'run_path_eq_trace_of_nice'
Previously the precondition was "m ≤ n - 1" (where 'm' is the number of
sprints in the run path and 'n' is the blocked list number) but it
should have been "m - 1 ≤ n". This difference turns out to be critical
because the case where m = n is required to complete the proof.

Specifically, the 'eaten' list upon which the endgame blocked list is
based must be the same as the list used to generate the last sprint of
endgame. If this is not the case, the upper bound on the number of
perimeter edges will not be tight enough to prove a contradiction.

Note that in a related mistake, 'make_block_list D p (n + 1)' was
commonly used when it should have been 'make_block_list D p (n - 1)'.
Fortunately, correcting these errors seems to have been mostly trivial.