Prove 'trace_perimeter_length_le_edges_card'
That is, the length of the perimeter is no greater than twice the size
of the blocked list, plus two. This is proven by demonstrating an
injective function from 'Fin P' to the edges of the blocked list region
and using 'region_edges_le' to establish an upper bound on the number
of edges (where 'P' is the length of the perimeter trace).